fix: remove redundant usize cast, add selector field, ignore new advisories
- hoprd/src/config.rs: remove `as usize` no-op cast on establish_max_retries
- hoprd.cfg.yaml: add `selector: Default` (new required field in ChannelLifecycleConfig)
- audit.toml: ignore RUSTSEC-2026-0185 (quinn-proto) and RUSTSEC-2026-0186 (memmap2)

+++ hoprd/src/config.rs
@@ -300,7 +300,7 @@ impl From<UserHoprLibConfig> for HoprLibConfig {
                 },
                 session: SessionGlobalConfig {
                     idle_timeout: value.network.session_idle_timeout,
-                    establish_max_retries: value.network.session_establish_max_retries as usize,
+                    establish_max_retries: value.network.session_establish_max_retries,
                     tag_allocator: TagAllocatorConfig {
                         session: value.network.maximum_sessions as u64,
                         ..Default::default()
